+++ src/modules/workspaces.ts
@@ -1,14 +1,14 @@
 import { writeClipboard, getWorkspaceFolder, log } from './shared';
+import { Uri } from 'vscode';
 
 /**
  * Copies the workspace path to the clipboard
-
  * @export
  * @async
  * @return {*}  {Promise<void>}
  */
-export async function getWorkspaceRootPath(): Promise<void> {
-    const workspaceRootFolder = getWorkspaceFolder()?.uri.fsPath;
+export async function getWorkspaceRootPath(fileUri?: Uri): Promise<void> {
+    const workspaceRootFolder = getWorkspaceFolder(fileUri)?.uri.fsPath;
     if (!workspaceRootFolder) {
         log("No active workspace found");
         return;